\name{WinPerfCounter.readcsv}
\alias{WinPerfCounter.readcsv}
\title{WinPerfCounters.csv}
\usage{
    WinPerfCounter.readcsv(filename)
}
\description{
    Reading Performance counter CSV file. return Dataframe.
}
\arguments{
    \item{filename}{filename}
}
\details{
    read the file.
}
\value{
    dataframe from csv.
}
\examples{
    
}
